+++ android/app/build.gradle
@@ -58,6 +58,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.6.2'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.2'
     implementation 'androidx.annotation:annotation:1.7.0'
+    implementation 'androidx.core:core-ktx:+'
     testImplementation 'junit:junit:4.13.2'
     implementation "com.kakao.sdk:v2-user:2.15.0"
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
